feat: add BetterStack uptime provider - refactor types to client
